Add SMD keepout zone to accommodate Pico 2's TP7
The Pico 2 is here!

TP7 declared "do not use", so I'm not planning to make a separate fleet
of footprints specifically for the Pico 2 yet.

Also, the geometry is almost identical to the original Pico
(only differences being small shifts of 2 rendered component positions
& rendering TP7), so I am hesitant to develop 3D models specifically
for the Pico 2, as it will dramatically increase library size for
minimal benefit.

<https://datasheets.raspberrypi.com/pico/pico-2-datasheet.pdf#_surface_mount_footprint>