Use more specific default footprints when possible
This way WYSIWYG in the render by default when selecting the
`RaspberryPi_Pico_W` schematic symbol.

Also favours conventional `_HandSolder` footprints as default when
possible, due to the Pico commonly being soldered by hand in hobbyist
settings.

Maintains `_Common_Unspecified` as the default for
`RaspberryPi_Pico` due to its more ambiguous name and likely
first-choice as a symbol for newer users unfamiliar with footprint
associations, providing them with a broad and non-specific default.